refactor: extract getClientIp to separate module for testability
Move getClientIp function from [id].ts to its own module to avoid
astro:content import resolution issues in Vitest.
